image/requirements.txt
Pillow==10.1.0
python-logging-loki==0.3.1
